c_symbols.binary -> c_symbols.find.
